+++ custom_components/device_emulator/time.py
@@ -34,7 +34,11 @@ class FakeTime(FakeEntityMixin, TimeEntity, RestoreEntity):
         self._attr_name = component.label
         self._attr_unique_id = f"{component.id}_time"
         self._attr_device_info = device_info_for(component.entry)
-        self._attr_native_value = dt_util.now().time().replace(microsecond=0)
+        self._attr_native_value = (
+            time.fromisoformat(component.initial)
+            if component.initial is not None
+            else dt_util.now().time().replace(microsecond=0)
+        )
 
     async def async_added_to_hass(self) -> None:
         await super().async_added_to_hass()